Improve QuickClip accessibility
