+++ java/res/values/strings-talkback-descriptions.xml
@@ -189,4 +189,11 @@
     <string name="action_menu_opened">%1$s mode</string>
     <!-- Spoken description when closing an action menu, returning back to main keyboard -->
     <string name="action_menu_closed">%1$s closed</string>
+
+
+    <string name="quick_clip_full_string">Quick clipboard item: “%s”. Double-tap to paste. Type anything to dismiss.</string>
+    <string name="quick_clip_numeric_code">Quick clip of an extracted numeric code: “%s”. Double-tap to paste. Type anything to dismiss.</string>
+    <string name="quick_clip_email_address">Quick clip of an extracted email address: “%s”. Double-tap to paste. Type anything to dismiss.</string>
+    <string name="quick_clip_url">Quick clip of an extracted URL: “%s”. Double-tap to paste. Type anything to dismiss.</string>
+    <string name="quick_clip_image">Quick image clipboard item. Double-tap to paste image. Type anything to dismiss.</string>
 </resources>
